requirements.txt
attrs==20.3.0
aws-cdk.alexa-ask==1.103.0
aws-cdk.assets==1.103.0
aws-cdk.aws-accessanalyzer==1.103.0
aws-cdk.aws-acmpca==1.103.0
aws-cdk.aws-amazonmq==1.103.0
aws-cdk.aws-amplify==1.103.0
aws-cdk.aws-apigateway==1.103.0
aws-cdk.aws-apigatewayv2==1.103.0
aws-cdk.aws-appconfig==1.103.0
aws-cdk.aws-appflow==1.103.0
aws-cdk.aws-appintegrations==1.103.0
aws-cdk.aws-applicationautoscaling==1.103.0
aws-cdk.aws-applicationinsights==1.103.0
aws-cdk.aws-appmesh==1.103.0
aws-cdk.aws-appstream==1.103.0
aws-cdk.aws-appsync==1.103.0
aws-cdk.aws-athena==1.103.0
aws-cdk.aws-auditmanager==1.103.0
aws-cdk.aws-autoscaling==1.103.0
aws-cdk.aws-autoscaling-common==1.103.0
aws-cdk.aws-autoscaling-hooktargets==1.103.0
aws-cdk.aws-autoscalingplans==1.103.0
aws-cdk.aws-backup==1.103.0
aws-cdk.aws-batch==1.103.0
aws-cdk.aws-budgets==1.103.0
aws-cdk.aws-cassandra==1.103.0
aws-cdk.aws-ce==1.103.0
aws-cdk.aws-certificatemanager==1.103.0
aws-cdk.aws-chatbot==1.103.0
aws-cdk.aws-cloud9==1.103.0
aws-cdk.aws-cloudformation==1.103.0
aws-cdk.aws-cloudfront==1.103.0
aws-cdk.aws-cloudtrail==1.103.0
aws-cdk.aws-cloudwatch==1.103.0
aws-cdk.aws-codeartifact==1.103.0
aws-cdk.aws-codebuild==1.103.0
aws-cdk.aws-codecommit==1.103.0
aws-cdk.aws-codedeploy==1.103.0
aws-cdk.aws-codeguruprofiler==1.103.0
aws-cdk.aws-codegurureviewer==1.103.0
aws-cdk.aws-codepipeline==1.103.0
aws-cdk.aws-codepipeline-actions==1.103.0
aws-cdk.aws-codestar==1.103.0
aws-cdk.aws-codestarconnections==1.103.0
aws-cdk.aws-codestarnotifications==1.103.0
aws-cdk.aws-cognito==1.103.0
aws-cdk.aws-config==1.103.0
aws-cdk.aws-customerprofiles==1.103.0
aws-cdk.aws-databrew==1.103.0
aws-cdk.aws-datapipeline==1.103.0
aws-cdk.aws-datasync==1.103.0
aws-cdk.aws-dax==1.103.0
aws-cdk.aws-detective==1.103.0
aws-cdk.aws-devopsguru==1.103.0
aws-cdk.aws-directoryservice==1.103.0
aws-cdk.aws-dlm==1.103.0
aws-cdk.aws-dms==1.103.0
aws-cdk.aws-docdb==1.103.0
aws-cdk.aws-dynamodb==1.103.0
aws-cdk.aws-ec2==1.103.0
aws-cdk.aws-ecr==1.103.0
aws-cdk.aws-ecr-assets==1.103.0
aws-cdk.aws-ecs==1.103.0
aws-cdk.aws-efs==1.103.0
aws-cdk.aws-eks==1.103.0
aws-cdk.aws-elasticache==1.103.0
aws-cdk.aws-elasticbeanstalk==1.103.0
aws-cdk.aws-elasticloadbalancing==1.103.0
aws-cdk.aws-elasticloadbalancingv2==1.103.0
aws-cdk.aws-elasticsearch==1.103.0
aws-cdk.aws-emr==1.103.0
aws-cdk.aws-emrcontainers==1.103.0
aws-cdk.aws-events==1.103.0
aws-cdk.aws-events-targets==1.103.0
aws-cdk.aws-eventschemas==1.103.0
aws-cdk.aws-fis==1.103.0
aws-cdk.aws-fms==1.103.0
aws-cdk.aws-fsx==1.103.0
aws-cdk.aws-gamelift==1.103.0
aws-cdk.aws-globalaccelerator==1.103.0
aws-cdk.aws-glue==1.103.0
aws-cdk.aws-greengrass==1.103.0
aws-cdk.aws-greengrassv2==1.103.0
aws-cdk.aws-groundstation==1.103.0
aws-cdk.aws-guardduty==1.103.0
aws-cdk.aws-iam==1.103.0
aws-cdk.aws-imagebuilder==1.103.0
aws-cdk.aws-inspector==1.103.0
aws-cdk.aws-iot==1.103.0
aws-cdk.aws-iot1click==1.103.0
aws-cdk.aws-iotanalytics==1.103.0
aws-cdk.aws-iotevents==1.103.0
aws-cdk.aws-iotfleethub==1.103.0
aws-cdk.aws-iotsitewise==1.103.0
aws-cdk.aws-iotthingsgraph==1.103.0
aws-cdk.aws-iotwireless==1.103.0
aws-cdk.aws-ivs==1.103.0
aws-cdk.aws-kendra==1.103.0
aws-cdk.aws-kinesis==1.103.0
aws-cdk.aws-kinesisanalytics==1.103.0
aws-cdk.aws-kinesisfirehose==1.103.0
aws-cdk.aws-kms==1.103.0
aws-cdk.aws-lakeformation==1.103.0
aws-cdk.aws-lambda==1.103.0
aws-cdk.aws-lambda-event-sources==1.103.0
aws-cdk.aws-licensemanager==1.103.0
aws-cdk.aws-logs==1.103.0
aws-cdk.aws-lookoutmetrics==1.103.0
aws-cdk.aws-lookoutvision==1.103.0
aws-cdk.aws-macie==1.103.0
aws-cdk.aws-managedblockchain==1.103.0
aws-cdk.aws-mediaconnect==1.103.0
aws-cdk.aws-mediaconvert==1.103.0
aws-cdk.aws-medialive==1.103.0
aws-cdk.aws-mediapackage==1.103.0
aws-cdk.aws-mediastore==1.103.0
aws-cdk.aws-msk==1.103.0
aws-cdk.aws-mwaa==1.103.0
aws-cdk.aws-neptune==1.103.0
aws-cdk.aws-networkfirewall==1.103.0
aws-cdk.aws-networkmanager==1.103.0
aws-cdk.aws-nimblestudio==1.103.0
aws-cdk.aws-opsworks==1.103.0
aws-cdk.aws-opsworkscm==1.103.0
aws-cdk.aws-pinpoint==1.103.0
aws-cdk.aws-pinpointemail==1.103.0
aws-cdk.aws-qldb==1.103.0
aws-cdk.aws-quicksight==1.103.0
aws-cdk.aws-ram==1.103.0
aws-cdk.aws-rds==1.103.0
aws-cdk.aws-redshift==1.103.0
aws-cdk.aws-resourcegroups==1.103.0
aws-cdk.aws-robomaker==1.103.0
aws-cdk.aws-route53==1.103.0
aws-cdk.aws-route53-targets==1.103.0
aws-cdk.aws-route53resolver==1.103.0
aws-cdk.aws-s3==1.103.0
aws-cdk.aws-s3-assets==1.103.0
aws-cdk.aws-s3-deployment==1.103.0
aws-cdk.aws-s3-notifications==1.103.0
aws-cdk.aws-s3objectlambda==1.103.0
aws-cdk.aws-s3outposts==1.103.0
aws-cdk.aws-sagemaker==1.103.0
aws-cdk.aws-sam==1.103.0
aws-cdk.aws-sdb==1.103.0
aws-cdk.aws-secretsmanager==1.103.0
aws-cdk.aws-securityhub==1.103.0
aws-cdk.aws-servicecatalog==1.103.0
aws-cdk.aws-servicecatalogappregistry==1.103.0
aws-cdk.aws-servicediscovery==1.103.0
aws-cdk.aws-ses==1.103.0
aws-cdk.aws-signer==1.103.0
aws-cdk.aws-sns==1.103.0
aws-cdk.aws-sns-subscriptions==1.103.0
aws-cdk.aws-sqs==1.103.0
aws-cdk.aws-ssm==1.103.0
aws-cdk.aws-sso==1.103.0
aws-cdk.aws-stepfunctions==1.103.0
aws-cdk.aws-synthetics==1.103.0
aws-cdk.aws-timestream==1.103.0
aws-cdk.aws-transfer==1.103.0
aws-cdk.aws-waf==1.103.0
aws-cdk.aws-wafregional==1.103.0
aws-cdk.aws-wafv2==1.103.0
aws-cdk.aws-workspaces==1.103.0
aws-cdk.cloud-assembly-schema==1.103.0
aws-cdk.cloudformation-include==1.103.0
aws-cdk.core==1.103.0
aws-cdk.custom-resources==1.103.0
aws-cdk.cx-api==1.103.0
aws-cdk.lambda-layer-awscli==1.103.0
aws-cdk.lambda-layer-kubectl==1.103.0
aws-cdk.region-info==1.103.0
boto3==1.17.72
botocore==1.20.72
#cattrs==1.5.0
#constructs==3.3.75
#jmespath==0.10.0
#jsii==1.29.0
#publication==0.0.3
#python-dateutil==2.8.1
#s3transfer==0.4.2
#six==1.16.0
#typing-extensions==3.10.0.0
#urllib3==1.26.4